check-downgrades: looks at module changes and disallows downgrades
Adds a command that will look at a git repository (and optional commit
reference) to see if any modules were downgraded. If any downgrades
happen, it must be explicitly allowed by a header in the commit
message, like

    Downgrade: <some/module>

or it will fail with an error and unique exit status.

Regardless of if it passes or fails, it outputs a summary of all
of the changes to the module versions: upgrades, downgrades, additions
and removals.

Maybe this can be a starting point for more analysis of our modules
in the future.

Change-Id: I39f3182281652d9037fb85520fe63d77e0d3307a