ci: fix aws-cli not being installed in ci image
first issue is that the install-awscli.sh script is not correctly
checking for errors, so we were still seeing successful builds of
the ci image, even when the aws-cli install was failing.

secondly, the aws-cli v1 no longer installs if it detects Python 2
installed on the system, which started occuring 15 July 2021.
Switching to installing the v2 client works fine, as it contains
its own interpreter and doesn't rely on installed packages.

attempting to remove Python 2 or set default to Python 3 may be a
rabbit hole, so rather than go down that path I think upgrading to
the latest aws-cli version is better. AWS recommends everyone upgrade
to this, and it works fine with our gateways.

Change-Id: I8c6776b3e6ec29a847a66e6dff070e75852ba7e8