Jenkinsfile: build and push ci image for main commits
currently in gateway-mt and other places, we docker build the ci
image on every build. In order to make builds more efficient, this
change will build and push the latest ci image to Docker Hub
whenever the ci main branch is pushed. This allows us to pull the
latest image instead of building it, which should improve build
times, as well as reduce load on the Jenkins agents.

Change-Id: I55d6d6131b02cc4248b37aad316b172541e637bd