check-downgrades: only alert on direct dependencies
if the dependency is not listed in `go mod why -m`, then
it only is in the graph for the module requirements
and not for any packages it provides. we don't care if
the module was downgraded in that case, though, because
any changes it could have caused are already being
checked.

Change-Id: I17ad4335735aa0638c7cfbcef9c1bf7b3cfe6579