fix(docker): support both amd64 and arm64 ci images
Many engineers develop on a Mac M1 chipset (arm64). While we can
virtualize many of the commands on this machine, developers pay a
performance cost due to the virtualization. As a result, we see
longer test times, linting times, etc. By supporting an arm64 CI
image, we drastically reduce the amount of time engineers spend
waiting for linting and tests to complete.

Change-Id: I1951511f52cbcdceb5f68f7a08384d59f3b456c5