check-downgrades: always run from git root
it was expected that users would pass -gitdir
to point at the root of the git repo. that leads
to hard to diagnose failures when run inside of
subdirectories of git repos where it would compare
different module lists for downgrades. instead
assume the user has run the command inside of
a git directory and always use the repo root.
we check all nested modules anyway, so there is
no worry about missing anything.

Change-Id: I864975dcad2b9befc7b6dd34e12695a551c46e8d