check-callsize: linter for large calls
Having large or a lot of arguments requires a lot of data to be
pushed onto the stack, which increases register pressure and
also makes calls themselves slower.

Note: This doesn't necessarily mean that replacing them with a
a single pointer options struct is more performant. However,
it can show potential issues.

Change-Id: Id3182bb9fdb1ef9094d2e74518567694b13a9e9e