images: bump Go 1.20.6 and misc. fixes
Bumping nodejs/npm in the same commit, because it seems the old script
isn't working as it used to.

Adjusting postgres to use the correct distro.

Removing `npx install playwright`, because that needs to be executed
from the context of the application and not globally.

Using `apt install python3-boto3` instead of pip3, because that's what
python wanted.

Also fix a bug in check-deferloop.

Change-Id: Ic7f5356a9be9ec2aa7b8a52beecd06226be19402