automerge: over-approximate query
the initial query just needs to ensure that it sees
the full potential set of changes it may want to take
an action on. the only downside to selecting more
changes is that it may be less aggressive about rebasing
due to some changes already being in a building state.
it will only try to merge changes that match the CanMerge
predicate which checks for submittable, etc. so there's
no need to encode that in the query.

Change-Id: I8a1efdaf09a9f6fa4e6cb6aa1c610d337b5a7dee